Add components for EducationalExperience
I'm not 100% sure what I was doing wrong here. If I remember correctly, I forgot to bind the handleSubmission method which caused an error with setState. This was particularly confusing but once I saw it, I was like "a-duh".

I also forgot to change the prop references in the display of content (not the forms) which was once everything else seemed to be working.

So, a couple lessons to learn here...